Refactored with LikeService interface and comments
